+++ vault/resource_kmip_secret_scope.go
@@ -28,7 +28,7 @@ func kmipSecretScopeResource() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				Description:  "Path where KMIP backend is mounted",
-				ValidateFunc: validateNoLeadingTrailingSlashes,
+				ValidateFunc: provider.ValidateNoLeadingTrailingSlashes,
 			},
 			"scope": {
 				Type:        schema.TypeString,
